fix: refactor to manage the components
