+++ client/components/Waiting.jsx
