+++ client/components/round/QuestionImage.jsx
